Shave a pointless timezone yak.
Compute the offset in hours of the host system and use it to set the
timezone of the vagrant guest.

Still isn't quite right because the offset may change, eg. daylight
savings, but you'll just have to run `vagrant provision` twice a year.

I don't want to deal with installing the `tzinfo` gem to run vagrant.